+++ packages/interface/src/index.ts
@@ -14,7 +14,7 @@
  * ```
  */
 
-import type { Connection, NewStreamOptions } from './connection.js'
+import type { Connection, NewStreamOptions, NewStreamProgressEvents } from './connection.js'
 import type { ContentRouting } from './content-routing.js'
 import type { Ed25519PublicKey, PublicKey, RSAPublicKey, Secp256k1PublicKey } from './keys.js'
 import type { Metrics } from './metrics.js'
@@ -54,6 +54,11 @@ export interface SignedPeerRecord {
   addresses: Multiaddr[]
   seq: bigint
 }
+
+/**
+ * Things that libp2p can dial
+ */
+export type DialTarget = PeerId | Multiaddr | Multiaddr[]
 
 /**
  * A certificate that can be used to secure connections
@@ -478,6 +483,8 @@ export type TransportManagerDialProgressEvents =
   ProgressEvent<'transport-manager:selected-transport', string>
 
 export type OpenConnectionProgressEvents =
+  ProgressEvent<'connection:open', DialTarget> |
+  ProgressEvent<'connection:opened', Connection> |
   TransportManagerDialProgressEvents |
   ProgressEvent<'dial-queue:already-connected'> |
   ProgressEvent<'dial-queue:already-in-dial-queue'> |
@@ -486,14 +493,14 @@ export type OpenConnectionProgressEvents =
   ProgressEvent<'dial-queue:calculated-addresses', Address[]> |
   OutboundConnectionUpgradeEvents
 
-export interface DialOptions extends AbortOptions, ProgressOptions {
+export interface DialOptions extends AbortOptions, ProgressOptions<OpenConnectionProgressEvents> {
   /**
    * If true, open a new connection to the remote even if one already exists
    */
   force?: boolean
 }
 
-export interface DialProtocolOptions extends NewStreamOptions {
+export interface DialProtocolOptions extends Omit<DialOptions, 'onProgress'>, Omit<NewStreamOptions, 'onProgress'>, ProgressOptions<OpenConnectionProgressEvents | NewStreamProgressEvents> {
 
 }
 
@@ -678,7 +685,7 @@ export interface Libp2p<T extends ServiceMap = ServiceMap> extends Startable, Ty
    * await conn.close()
    * ```
    */
-  dial(peer: PeerId | Multiaddr | Multiaddr[], options?: DialOptions): Promise<Connection>
+  dial(peer: DialTarget, options?: DialOptions): Promise<Connection>
 
   /**
    * Dials to the provided peer and tries to handshake with the given protocols in order.
@@ -696,7 +703,7 @@ export interface Libp2p<T extends ServiceMap = ServiceMap> extends Startable, Ty
    * pipe([1, 2, 3], stream, consume)
    * ```
    */
-  dialProtocol(peer: PeerId | Multiaddr | Multiaddr[], protocols: string | string[], options?: DialProtocolOptions): Promise<Stream>
+  dialProtocol(peer: DialTarget, protocols: string | string[], options?: DialProtocolOptions): Promise<Stream>
 
   /**
    * Attempts to gracefully close an open connection to the given peer. If the