+++ packages/connection-encrypter-tls/CHANGELOG.md
@@ -6,6 +6,20 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [3.0.15](https://github.com/libp2p/js-libp2p/compare/tls-v3.0.14...tls-v3.0.15) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+    * @libp2p/peer-id bumped from ^6.0.5 to ^6.0.6
+    * @libp2p/utils bumped from ^7.0.14 to ^7.0.15
+  * devDependencies
+    * @libp2p/logger bumped from ^6.2.3 to ^6.2.4
+
 ## [3.0.14](https://github.com/libp2p/js-libp2p/compare/tls-v3.0.13...tls-v3.0.14) (2026-03-27)