+++ packages/libp2p-daemon-protocol/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 * bump aegir from 46.0.5 to 47.0.21 ([#343](https://github.com/libp2p/js-libp2p-daemon/issues/343)) ([704c22f](https://github.com/libp2p/js-libp2p-daemon/commit/704c22f102362c6036642a73979d262e1214baa5))
 
+## [8.0.5](https://github.com/libp2p/js-libp2p/compare/daemon-protocol-v8.0.4...daemon-protocol-v8.0.5) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+
 ## [8.0.4](https://github.com/libp2p/js-libp2p/compare/daemon-protocol-v8.0.3...daemon-protocol-v8.0.4) (2026-03-27)